+++ src/Security/Security/ChangeLog.md
@@ -19,7 +19,7 @@
 -->
 
 ## Upcoming Release
-
+* Added new cmdlet: `Get-AzSecuritySolution`
 ## Version 1.1.1
 * Updated Security .NET SDK package reference to version 3.0.0