+++ src/Security/Security/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+
+## Version 1.2.0
 * Added new cmdlets for security Automations API
 
 ## Version 1.1.1